rag-service/requirements.txt
# RAG Service Dependencies
# ========================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9  # For file uploads

# Configuration
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1

# LangChain Core
langchain==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-community==0.3.0

# LLM Providers
langchain-groq==0.2.0
langchain-huggingface==0.1.0

# Embeddings & Vector Store
sentence-transformers==3.1.0
faiss-cpu==1.8.0

# Utilities
numpy>=1.26.0
torch>=2.0.0  # Required by sentence-transformers
